feat: NewAddCouponSection, usdAddCoupon 분리
